+++ tests/test_controller.py
@@ -1,5 +1,15 @@
 import pytest
-from src.controller import app
+import sys
+import os
+
+# Get the current directory (where this test file is located)
+current_dir = os.path.dirname(os.path.abspath(__file__))
+
+# Add the parent directory to sys.path (the src directory)
+parent_dir = os.path.join(current_dir, "..")
+sys.path.insert(0, parent_dir)
+
+from src.run import app
 
 @pytest.fixture
 def client():
